+++ hyprxa/auth/route.py
@@ -1,5 +1,8 @@
+import secrets
+
 from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordRequestForm
+from jose import jwt
 
 from hyprxa.auth.models import Token, TokenHandler
 from hyprxa.auth.protocols import AuthenticationClient
@@ -22,3 +25,10 @@ async def token(
         status_code=status.HTTP_401_UNAUTHORIZED,
         detail="Incorrect username or password"
     )
+
+
+async def debug_token(form: OAuth2PasswordRequestForm = Depends()) -> Token:
+    """Application is in debug mode. Token has no effect."""
+    claims = {"sub": form.username}
+    access_token = jwt.encode(claims=claims, key=secrets.token_hex(32), algorithm="HS256")
+    return Token(access_token=access_token, token_type="bearer")